+++ ChangeLog.md
@@ -14,6 +14,9 @@ IMPROVEMENTS
 
 BUG FIXES
 
+- fix `--restore-database-mapping` with special characters in source table, fix
+  reopened [820](https://github.com/Altinity/clickhouse-backup/issues/820#issuecomment-2675628282),
+  thanks @IvaskevychYuriy
 - add workaround for `metdata_path` field change behavior in 25.1+, look details
   in https://github.com/ClickHouse/ClickHouse/issues/76546,
   fix [1093](https://github.com/Altinity/clickhouse-backup/issues/1093)